[BugFix] Timing Fix
Added some constants that allow for easy adjustment of fps and in game timing logic. Also did some minor refactoring throughout the program.
