+++ Context.py
@@ -1,6 +1,8 @@
 import math
 import weather_data.soldata as soldata
 import random
+import Plant
+import GameInfo
 
 from enum import Enum
  
@@ -8,147 +10,6 @@ class GameState(Enum):
     TITLE = 1
     HELP  = 2
     GAME  = 3
-
-
-class Crop:
-    def __init__(self):
-        self.name = "Parent crop."
-        self.quantity = 0           # All start at 0 quantity.
-        self.minGoodTemp = 10       # Dummy value.
-        self.safeUvLevels = {"Very_High": False, "High": False, "Moderate": False, "Low":False}          # Dummy value.
-        self.sellValue = 0              # Dummy value.
-        self.buyValue = 0      # Dummy value.
-        self.foodPerHourPerPlant = 0        # Dummy value
-        self.spriteFile = ""
-        self.spriteCoords = []
-        self.thresh = 0
-        self.to_show = 0
-
-    # Increases the number of plants of this crop by one. It is the implementing function's responsibility to check that there are sufficient funds.
-    def addPlant(self, x1, y1, x2, y2, dead_x1, dead_y1, dead_x2, dead_y2):
-        self.quantity += 1
-        # Possibly increase price per plant on purchase?
-
-        coords = [random.randint(x1, x2), random.randint(y1, y2)]
-
-        while dead_x1 <= coords[0] <= dead_x2 and dead_y1 <= coords[1] <= dead_y2:
-            coords = [random.randint(x1, x2), random.randint(y1, y2)]
-
-        self.to_show += 1
-        self.spriteCoords.append(coords)
-
-    # Decreases the number of plants, if possible. Returns true on a success and false on a failure.
-    def sellPlant(self):
-        if self.quantity >= 1:
-            self.quantity -= 1
-            self.spriteCoords.pop()
-            return True
-        else:
-            return False
-    
-    # Returns the total food value generated by this plant per tick.
-    def getFoodPerTick(self, temp, uvLevel):
-        valuePercent = 1.0
-
-        if temp < self.minGoodTemp:
-            valuePercent -= .5
-        
-        if not self.safeUvLevels[uvLevel]:
-            valuePercent -= .5
-
-
-        return self.quantity * self.foodPerHourPerPlant * valuePercent / Context.ticksPerHour
-
-    
-    # Returns a brief string description for the crop.
-    def getUvTolerance(self):
-        uvTolerance = ""
-        if self.safeUvLevels["High"]:
-            uvTolerance += "High"
-        elif self.safeUvLevels["Moderate"]:
-            uvTolerance += "Moderate"
-        elif self.safeUvLevels["Low"]:
-            uvTolerance += "Low"
-
-        return f"UV Tolerance: {uvTolerance}"
-    
-class generic(Crop):
-
-    def __init__(self):
-        self.name = "Super Tomato"
-        self.quantity = 0           # All start at 0 quantity.
-        self.minGoodTemp = -50      
-        self.safeUvLevels = {"Very_High": False, "High": False, "Moderate": True, "Low":True}     
-        self.sellValue = 10              
-        self.buyValue = 20     
-        self.foodPerHourPerPlant = 5
-        self.spriteFile = "assets\\green_plant.png"
-        self.spriteCoords = []   
-        self.thresh = 0   
-        self.to_show = 0  
-
-
-class uvResistant(Crop):
-
-    def __init__(self):
-        self.name = "Resistant Radishes"
-        self.quantity = 0           # All start at 0 quantity.
-        self.minGoodTemp = -50      
-        self.safeUvLevels = {"Very_High": False, "High": True, "Moderate": True, "Low":True}       
-        self.sellValue = 15             
-        self.buyValue = 25     
-        self.foodPerHourPerPlant = 3.5        
-        self.spriteFile = "assets\\red_plant.png"
-        self.spriteCoords = []   
-        self.thresh = 1000 
-        self.to_show = 0  
-
-class coldResistant(Crop):
-
-    def __init__(self):
-        self.name = "Freeze Free Lettuce"
-        self.quantity = 0           # All start at 0 quantity.
-        self.minGoodTemp = -90      
-        self.safeUvLevels = {"Very_High": False, "High": False, "Moderate": True, "Low":True}     
-        self.sellValue = 15              
-        self.buyValue = 25     
-        self.foodPerHourPerPlant = 2
-        self.spriteFile = "assets\\blue_plant.png"
-        self.spriteCoords = []      
-        self.thresh = 500  
-        self.to_show = 0       
-
-class hybrid(Crop):
-
-    def __init__(self):
-        self.name = "Uber Tubers"
-        self.quantity = 0           # All start at 0 quantity.
-        self.minGoodTemp = -30      
-        self.safeUvLevels = {"Very_High": False, "High": False, "Moderate": True, "Low":True}      
-        self.sellValue = 15              
-        self.buyValue = 30     
-        self.foodPerHourPerPlant = 7.5 
-        self.spriteFile = "assets\\potato.png"
-        self.spriteCoords = []   
-        self.thresh = 5000 
-        self.to_show = 0      
-
-class cashcow(Crop):
-
-    def __init__(self):
-
-        self.name = "Super Sweet Strawberries"
-
-        self.quantity = 0           # All start at 0 quantity.
-        self.minGoodTemp = -10      
-        self.safeUvLevels = {"Very_High": False, "High": False, "Moderate": False, "Low":True}     
-        self.sellValue = 15              
-        self.buyValue = 30     
-        self.foodPerHourPerPlant = 25    
-        self.spriteFile = "assets\\cash_cow.png"
-        self.spriteCoords = [] 
-        self.thresh = 10000 
-        self.to_show = 0   
 
 class Context:
     #constants
@@ -162,16 +23,8 @@ class Context:
 
     ADTL_PERSON_THRESH = 500
 
-    # Timing Constants
-    tick = .1
-    ticksPerMinute = 1
-    ticksPerHour = 120
-    ticksPerDay = 2880
-    minuteIncrement = 10
+    DAILY_FOOD_INTAKE_PER_PERSON = 10
 
-    minPerHour = 60
-    hourPerSol = 24
-    solPerYear = 668 
 
     # defining colors
     white = (255, 255, 255)
@@ -190,6 +43,7 @@ class Context:
         self.population = 0
         self.gainedFood = 0
         self.consumedFood = 0
+        self.totalConsumedFood = self.consumedFood
 
         # define click regions
         self.click_regions = []
@@ -214,11 +68,11 @@ class Context:
 
             # define crop type objects. store as a list for easy iteration
         self.crp = [
-            generic(),    
-            coldResistant(),
-            uvResistant(),
-            hybrid(),
-            cashcow()
+            Plant.Generic(),    
+            Plant.ColdResistant(),
+            Plant.UVResistant(),
+            Plant.Hybrid(),
+            Plant.CashCow()
         ]
 
         self.crops = []
@@ -267,32 +121,37 @@ class Context:
 
     def update_people(self):
         
-        if self.totalfood >= self.PEOPLE_THRESH:
-            self.population = round((self.totalfood - self.PEOPLE_THRESH)/self.ADTL_PERSON_THRESH)
+        availableFood = self.totalfood - self.totalConsumedFood - self.PEOPLE_THRESH
+
+        if availableFood >= 0:
+            self.population = round((availableFood)/self.ADTL_PERSON_THRESH)
             
             if self.first_person == "Shown":
                 return
             elif self.first_person == "Insufficent-Food" and self.population >= 1:
                 self.first_person = "To-Show"
+                
+        elif self.population > 0:
+            self.population = 0
             
             
     
     
     def next_tick(self):
-        self.doTickUpdate()
+        self.foodTickUpdate()
         self.update_crops()
         self.update_people()
         self.tickCounter += 1
 
-        if self.tickCounter >= Context.ticksPerMinute:
+        if self.tickCounter >= GameInfo.TICKS_PER_MINUTE:
             self.tickCounter = 0
             self.minNum += 1
 
-        if self.minNum >= Context.minPerHour:
+        if self.minNum >= GameInfo.MIN_PER_HOUR:
             self.minNum = 0
             self.hourNum += 1
 
-        if self.hourNum >= Context.hourPerSol:
+        if self.hourNum >= GameInfo.HOUR_PER_SOL:
             self.hourNum = 0
             self.increment_sol()
 
@@ -301,7 +160,7 @@ class Context:
         max_temp = self.current_env['max_temp']
         if minutes == None:
             minutes = self.minNum + self.hourNum * 60
-        minutesInDay = 1440
+        minutesInSol = GameInfo.MIN_PER_HOUR * GameInfo.HOUR_PER_SOL
         temp_range = max_temp - min_temp
 
 
@@ -310,21 +169,19 @@ class Context:
         #in the hopes we can somewhat smoothly catch up or wait for the sinusoidal approximation
         #when switching between days
         if minutes < linearthreshold:
-            cosapprox = min_temp + .5 * temp_range + .5 * temp_range * math.cos((linearthreshold + self.current_env['sunrise'] + 60) * 2 * math.pi / minutesInDay)
+            cosapprox = min_temp + .5 * temp_range + .5 * temp_range * math.cos((linearthreshold + self.current_env['sunrise'] + 60) * 2 * math.pi / minutesInSol)
             m = (cosapprox - self.prevTemp) / linearthreshold
             b = self.prevTemp
             return m * minutes + b
         #if we're an hour in we'll do a sinusoidal approximation
-        return min_temp + .5 * temp_range + .5 * temp_range * math.cos((minutes + self.current_env['sunrise'] + 60) * 2 * math.pi / minutesInDay)
-#        if 60 * self.hourNum + self.minNum < self.current_env['sunrise'] or 60 * self.hourNum + self.minNum > self.current_env['sunset']:
-#            return self.current_env['min_temp']
-#        return self.current_env['max_temp']
+        return min_temp + .5 * temp_range + .5 * temp_range * math.cos((minutes + self.current_env['sunrise'] + 60) * 2 * math.pi / minutesInSol)
     
     def get_pressure(self):
         return self.current_env['pressure']
     
     def get_uv(self):
-        if 60 * self.hourNum + self.minNum < self.current_env['sunrise'] or 60 * self.hourNum + self.minNum > self.current_env['sunset']:
+        timeInMinutes = GameInfo.MIN_PER_HOUR * self.hourNum + self.minNum
+        if  timeInMinutes < self.current_env['sunrise'] or timeInMinutes > self.current_env['sunset']:
             return "Low"
         return self.current_env['uv_index']
     
@@ -336,32 +193,34 @@ class Context:
         return self._format_time(self.current_env['sunset'])
     
     def get_production(self):
-        return self.gainedFood * self.ticksPerHour
+        return self.gainedFood * GameInfo.TICKS_PER_HOUR
     
     def get_consumption(self):
-        return self.consumedFood * self.ticksPerHour
+        return self.consumedFood * GameInfo.TICKS_PER_HOUR
     
     def get_net(self):
         return self.get_production() - self.get_consumption()
 
     def _format_time(self, min_since_midnight):
         hr = str(math.floor(min_since_midnight / 60))
-        min = str(math.floor(min_since_midnight % 60))
+        mins = str(math.floor(min_since_midnight % 60))
         if len(hr) < 2:
             hr = '0' + hr
-        if len(min) < 2:
-            min = '0' + min
-        return hr + ":" + min
+        if len(mins) < 2:
+            mins = '0' + mins
+        return hr + ":" + mins
     
-    def doTickUpdate(self):
+    def foodTickUpdate(self):
         self.gainedFood = 0
-        self.consumedFood = self.population * 3 /self.ticksPerDay * 100
+        self.consumedFood = self.population * Context.DAILY_FOOD_INTAKE_PER_PERSON / GameInfo.TICKS_PER_SOL * 100
         for crop in self.crops:
 
             self.gainedFood += crop.getFoodPerTick(self.get_temp(), self.get_uv())
 
-        self.food += self.gainedFood - self.consumedFood
-        self.totalfood += self.gainedFood - self.consumedFood
+        netFood = self.gainedFood - self.consumedFood
+        self.food += netFood
+        self.totalConsumedFood -= self.consumedFood
+        self.totalfood += self.gainedFood
 
     def _handle_click_event(self, event, mouse):
         
@@ -422,8 +281,6 @@ class Context:
         self.click_regions = []
 
     def append_click_region(self, x, y, width, height, str):
-      
-
         self.click_regions.append({'x':x, 'y':y, 'width':width, 'height':height, 'desc': str})
     
     def increment_sol(self):
@@ -432,7 +289,7 @@ class Context:
         self.food -= self.consumedFood
 
         self.solNum += 1
-        if self.solNum > Context.solPerYear:
+        if self.solNum > GameInfo.SOL_PER_YEAR:
             self.solNum = 1
             self.yearNum += 1
         
@@ -459,6 +316,6 @@ class Context:
         self.current_env = soldata.getRandomizedSolData(self.solNum - 1, self.sol_data)
     
     def get_time_string(self):
-        return f"Year {self.yearNum}, Sol {self.solNum}        {str(self.hourNum).zfill(2)}:{str(self.minNum - self.minNum % Context.minuteIncrement).zfill(2)}"
+        return f"Year {self.yearNum}, Sol {self.solNum}        {str(self.hourNum).zfill(2)}:{str(self.minNum - self.minNum % GameInfo.MINUTE_INCREMENT).zfill(2)}"
